third_party/blink/web_tests/external/wpt/css/css-contain/container-queries/at-container-serialization.html

<!doctype html>
<title>CSS Container Queries: @container serialization</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-queries">
<link rel="help" href="https://drafts.csswg.org/cssom/#serialize-a-css-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style id="testSheet">
  @container (width=100px) {
    @container \!-name (inline-size > 200px    ) {
      #id { color: lime }
    }
    #id { color: green }
  }
</style>
<script>
  setup(() => assert_implements_container_queries());

  let rules = testSheet.sheet.cssRules;

  test(() => {
    assert_equals(rules.length, 1);
    assert_equals(rules[0].cssRules.length, 2);

    assert_equals(rules[0].conditionText, "(width = 100px)");
    assert_equals(rules[0].cssRules[0].conditionText, "(inline-size > 200px)");
  }, "Serialization of conditionText");

  test(() => {
    assert_equals(rules[0].cssRules[0].cssText, "@container \\!-name (inline-size > 200px) {\n  #id { color: lime; }\n}");
  }, "Serialization of inner @container rule");

  test(() => {
    assert_equals(rules[0].cssText, "@container (width = 100px) {\n  @container \\!-name (inline-size > 200px) {\n  #id { color: lime; }\n}\n  #id { color: green; }\n}");
  }, "Serialization of nested @container rule");

</script>
